+++ src/components/seller/index.ts
@@ -1,1 +1,4 @@
 export { default as SellerTabs } from "./SellerTabs";
+export { SellerStatCard } from "./SellerStatCard";
+export { SellerQuickActions } from "./SellerQuickActions";
+export { SellerRecentListings } from "./SellerRecentListings";